requirements.txt
dj-database-url==0.4.1
Django==1.10.4
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
behave-django==0.5.0
splinter==0.7.6
